refactor: split navigation sidebar into smaller components
