+++ frontend/src/component/layout/MainLayout/NavigationSidebar/IconRenderer.tsx
@@ -1,0 +1,79 @@
+import EmptyIcon from '@mui/icons-material/CheckBoxOutlineBlankOutlined';
+import type SvgIcon from '@mui/material/SvgIcon/SvgIcon';
+import ApplicationsIcon from '@mui/icons-material/AppsOutlined';
+import ContextFieldsIcon from '@mui/icons-material/AccountTreeOutlined';
+import FlagTypesIcon from '@mui/icons-material/OutlinedFlag';
+import IntegrationsIcon from '@mui/icons-material/IntegrationInstructionsOutlined';
+import SegmentsIcon from '@mui/icons-material/DonutLargeOutlined';
+import CustomStrategiesIcon from '@mui/icons-material/ExtensionOutlined';
+import TagTypesIcon from '@mui/icons-material/LabelImportantOutlined';
+import EnvironmentsIcon from '@mui/icons-material/CloudOutlined';
+import UsersIcon from '@mui/icons-material/GroupOutlined';
+import ServiceAccountIcon from '@mui/icons-material/SmartToyOutlined';
+import GroupsIcon from '@mui/icons-material/GroupsOutlined';
+import RoleIcon from '@mui/icons-material/AdminPanelSettingsOutlined';
+import ApiAccessIcon from '@mui/icons-material/KeyOutlined';
+import SingleSignOnIcon from '@mui/icons-material/AssignmentOutlined';
+import NetworkIcon from '@mui/icons-material/HubOutlined';
+import MaintenanceIcon from '@mui/icons-material/BuildOutlined';
+import BannersIcon from '@mui/icons-material/PhotoOutlined';
+import InstanceStatsIcon from '@mui/icons-material/QueryStatsOutlined';
+import LicenseIcon from '@mui/icons-material/ReceiptLongOutlined';
+import InstancePrivacyIcon from '@mui/icons-material/ShieldOutlined';
+import LoginHistoryIcon from '@mui/icons-material/HistoryOutlined';
+import CorsIcon from '@mui/icons-material/StorageOutlined';
+import BillingIcon from '@mui/icons-material/CreditCardOutlined';
+import EventLogIcon from '@mui/icons-material/EventNoteOutlined';
+import GitHubIcon from '@mui/icons-material/GitHub';
+import LibraryBooksIcon from '@mui/icons-material/LibraryBooks';
+import { ReactComponent as ProjectIcon } from 'assets/icons/projectIconSmall.svg';
+import type { FC } from 'react';
+import { styled } from '@mui/material';
+
+const icons: Record<string, typeof SvgIcon> = {
+    '/applications': ApplicationsIcon,
+    '/context': ContextFieldsIcon,
+    '/feature-toggle-type': FlagTypesIcon,
+    '/integrations': IntegrationsIcon,
+    '/segments': SegmentsIcon,
+    '/strategies': CustomStrategiesIcon,
+    '/tag-types': TagTypesIcon,
+    '/environments': EnvironmentsIcon,
+    '/admin/users': UsersIcon,
+    '/admin/service-accounts': ServiceAccountIcon,
+    '/admin/groups': GroupsIcon,
+    '/admin/roles': RoleIcon,
+    '/admin/api': ApiAccessIcon,
+    '/admin/auth': SingleSignOnIcon,
+    '/admin/network': NetworkIcon,
+    '/admin/maintenance': MaintenanceIcon,
+    '/admin/banners': BannersIcon,
+    '/admin/instance': InstanceStatsIcon,
+    '/admin/license': LicenseIcon,
+    '/admin/instance-privacy': InstancePrivacyIcon,
+    '/admin/logins': LoginHistoryIcon,
+    '/admin/cors': CorsIcon,
+    '/admin/billing': BillingIcon,
+    '/history': EventLogIcon,
+    GitHub: GitHubIcon,
+    Documentation: LibraryBooksIcon,
+};
+
+const findIcon = (key: string) => {
+    return icons[key] || EmptyIcon;
+};
+
+export const IconRenderer: FC<{ path: string }> = ({ path }) => {
+    const IconComponent = findIcon(path); // Fallback to 'default' if the type is not found
+
+    return <IconComponent />;
+};
+
+export const StyledProjectIcon = styled(ProjectIcon)(({ theme }) => ({
+    fill: theme.palette.neutral.main,
+    stroke: theme.palette.neutral.main,
+    // same as built-in icons
+    width: theme.spacing(3),
+    height: theme.spacing(3),
+    fontSize: theme.spacing(3),
+}));
