+++ frontend/src/component/layout/MainLayout/NavigationSidebar/NavigationMode.ts
@@ -1,0 +1,1 @@
+export type NavigationMode = 'mini' | 'full';
